+++ include/build_version.h
@@ -8,10 +8,10 @@
 #ifndef BUILD_VERSION_H
 #define BUILD_VERSION_H
 
-#define BUILD_NUMBER 511
-#define BUILD_TIMESTAMP "2025-12-11 19:10:01"
-#define BUILD_DATE "20251211"
-#define GIT_HASH "11e6948"
+#define BUILD_NUMBER 514
+#define BUILD_TIMESTAMP "2025-12-12 15:13:00"
+#define BUILD_DATE "20251212"
+#define GIT_HASH "4df52ed"
 #define GIT_BRANCH "main"
 
 #endif // BUILD_VERSION_H